Backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pymongo>=4.0
python-dotenv>=1.0
elevenlabs>=1.0
